# docs/swagger.yaml
definitions:
  github_com_luizweitz_go-api_internal_models.Error:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  github_com_luizweitz_go-api_internal_models.User:
    properties:
      age:
        type: integer
      city:
        description: can be null
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    required:
    - age
    - email
    - name
    type: object
info:
  contact:
    name: Luiz Weitz
    url: http://www.swagger.io/support
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Search All Users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.User'
            type: array
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
      summary: Search All Users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create User With The Given Input Data
      parameters:
      - description: Create user object
        in: body
        name: Input
        required: true
        schema:
          $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.User'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
        "409":
          description: 'Error Conflict '
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
      summary: Create User
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update Data User By ID With The Given Input Data
      parameters:
      - description: Update user object
        in: body
        name: Input
        required: true
        schema:
          $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
        "404":
          description: Record Not Found
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
      summary: Update User By ID
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User By ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
        "404":
          description: Record Not Found
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
      summary: Delete User By ID
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get User By ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.User'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
        "404":
          description: Eecord Not Found
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_luizweitz_go-api_internal_models.Error'
      summary: Search User By ID
      tags:
      - users
swagger: "2.0"